Move the frontend test servers to their own modules
